+++ libraries/session/src/main/java/androidx/media3/session/MediaConstants.java
@@ -497,6 +497,8 @@ public final class MediaConstants {
       "androidx.media3.session.SESSION_COMMAND_ON_CAPTIONING_ENABLED_CHANGED";
   /* package */ static final String SESSION_COMMAND_REQUEST_SESSION3_TOKEN =
       "androidx.media3.session.SESSION_COMMAND_REQUEST_SESSION3_TOKEN";
+  /* package */ static final String SESSION_COMMAND_MEDIA3_PLAY_REQUEST =
+      "androidx.media3.session.SESSION_COMMAND_MEDIA3_PLAY_REQUEST";
 
   /* package */ static final String ARGUMENT_CAPTIONING_ENABLED =
       "androidx.media3.session.ARGUMENT_CAPTIONING_ENABLED";